REC-204 Move user actions from preprocessor to preprocessor_common
