+++ environment.yml
@@ -41,7 +41,7 @@ dependencies:
     - pandas==1.4.2
     - pyarrow==10.0.1
     - pymongo==4.1.0
-    - pymongoarrow==0.6.2
+    - pymongoarrow==0.7.0
     - python-dateutil==2.8.2
     - python-dotenv==0.20.0
     - pytz==2022.1
