+++ requirements.txt
@@ -28,5 +28,5 @@ urllib3==1.26.9
 Werkzeug==2.1.2
 zipp==3.8.0
 flask-pymongo==2.3.0
-pymongoarrow==0.6.2
+pymongoarrow==0.7.0
 flake8==6.0.0
